End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { MessageService } from './../_services/message.service';
 import { Pagination } from './../_models/pagination';
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -16,26 +17,31 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages()
   }
 
-  loadMessages(){
+  loadMessages() {
     this.loading = true;
     this.messageService.getMessages(this.pageNumber, this.pageSize, this.container)
-      .subscribe(response =>{
+      .subscribe(response => {
         this.messages = response.result;
         this.pagination = response.pagination;
         this.loading = false;
       })
   }
 
-  deleteMessage(id: number){
-    this.messageService.deleteMessage(id).subscribe(()=>{
-      this.messages.splice(this.messages.findIndex(m=>m.id === id), 1);
-    })
+  deleteMessage(id: number) {
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone')
+      .subscribe(result => {
+        if (result) {
+          this.messageService.deleteMessage(id).subscribe(() => {
+            this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+          })
+        }
+      })
   }
 
   pageChanged(event: any) {
